move and rename modules
Up until now modules have been named so that they mirror the folder
structure of MSDN's documentation. The main reason for this was to
assist library users refering to Microsoft's documentation. Likewise,
each data object in the library was given its own module with the same
name. In practice this resulted in long, awkward names. The benefits to
mirroring Microsoft's documentation structure are not as great as I
originally thought.