internal refactoring
ServiceState now has an instance for Storable. Invalid DWORD values
marshaled in as a ServiceState will trigger a call to error.

ServiceType now has an instance for Storable. Invalid DWORD values
marshaled in as a ServiceState will trigger a call to error.

Internal function names changed to avoid conflicting with Win32-errors.