Move type definitions for callback functions
The underlying library uses callback functions in its operation. We
represent these in Haskell as real functions which are converted to
FunPtr objects internally. It makes more sense to define these near
functions that use them rather than a central Types module. With this
change we are able to remove the Types module altogether.